Minor refactoring. Ids should be longs.

+++ sdk/src/main/java/com/irewind/sdk/model/Tag.java
@@ -5,12 +5,12 @@ import com.google.gson.annotations.SerializedName;
 public class Tag extends BaseResponse {
 
     @SerializedName("tagId")
-    private Integer tagId;
+    private long tagId;
 
     @SerializedName("tagName")
     private String tagName;
 
-    public Integer getTagId() {
+    public long getTagId() {
         return tagId;
     }
